got rid of excess interface
