--- QuickFIXn/IAssumedSessionSet.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-
-namespace QuickFix
-{
-    internal interface IAssumedSessionSet
-    {
-        IEnumerable<SessionID> GetSessionIds();
-    }
-}
