+++ QuickFIXn/QuickFix.csproj
@@ -47,6 +47,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AcceptorSocketDescriptor.cs" />
     <Compile Include="CompositeLogFactory.cs" />
     <Compile Include="CompositeLog.cs" />
     <Compile Include="DataDictionary\IFieldMapSpec.cs" />
@@ -75,7 +76,6 @@
     <Compile Include="IMessageStore.cs" />
     <Compile Include="IMessageStoreFactory.cs" />
     <Compile Include="IResponder.cs" />
-    <Compile Include="IAssumedSessionSet.cs" />
     <Compile Include="MemoryStore.cs" />
     <Compile Include="MemoryStoreFactory.cs" />
     <Compile Include="MessageBuilder.cs" />
