+++ QuickFIXn/ThreadedSocketReactor.cs
@@ -35,7 +35,7 @@ namespace QuickFix
         private TcpListener tcpListener_;
         private SocketSettings socketSettings_;
         private QuickFix.Dictionary sessionDict_;
-        private readonly IAssumedSessionSet assumedSessionSet_;
+        private readonly AcceptorSocketDescriptor acceptorDescriptor_;
 
         #endregion
 
@@ -49,12 +49,12 @@ namespace QuickFix
         {
             
         }
-        internal ThreadedSocketReactor(IPEndPoint serverSocketEndPoint, SocketSettings socketSettings, QuickFix.Dictionary sessionDict, IAssumedSessionSet assumedSessionSet)
+        internal ThreadedSocketReactor(IPEndPoint serverSocketEndPoint, SocketSettings socketSettings, QuickFix.Dictionary sessionDict, AcceptorSocketDescriptor acceptorDescriptor)
         {
             socketSettings_ = socketSettings;
             tcpListener_ = new TcpListener(serverSocketEndPoint);
             sessionDict_ = sessionDict;
-            assumedSessionSet_ = assumedSessionSet;
+            acceptorDescriptor_ = acceptorDescriptor;
         }
 
         public void Start()
@@ -101,7 +101,7 @@ namespace QuickFix
                     TcpClient client = tcpListener_.AcceptTcpClient();
                     ApplySocketOptions(client, socketSettings_);
                     ClientHandlerThread t = 
-                        new ClientHandlerThread(client, nextClientId_++, sessionDict_, socketSettings_, assumedSessionSet_);
+                        new ClientHandlerThread(client, nextClientId_++, sessionDict_, socketSettings_, acceptorDescriptor_);
                     t.Exited += OnClientHandlerThreadExited;
                     lock (sync_)
                     {
